requirements.txt
streamlit==1.31.1

transformers==4.38.2
torch==2.2.1
sentence-transformers==2.6.1
faiss-cpu==1.7.4

langchain==0.1.13
openai==1.12.0

pandas==2.2.1
requests==2.31.0
tqdm==4.66.2
sqlalchemy==2.0.27
python-dotenv==1.0.1